requirements.txt
Pillow
PyMuPDF
fastapi
google-generativeai
openpyxl
pydantic
pymilvus
pytesseract
python-docx
python-multipart
sentence-transformers
tqdm
uvicorn
